Move resolveOutputFilePropertySpecs() to OutputUnpacker
